clean up EnvUtils by moving Android-specific stuff to AndroidEnvUtils
