+++ maestro-cli/src/main/java/maestro/cli/util/AndroidEnvUtils.kt
@@ -4,16 +4,56 @@ import maestro.cli.CliError
 import okio.buffer
 import okio.source
 import java.io.File
+import java.nio.file.Path
+import java.nio.file.Paths
 import java.util.concurrent.TimeUnit
 import java.util.concurrent.TimeoutException
 
 object AndroidEnvUtils {
 
+    private val androidHome: String?
+        get() {
+            return System.getenv("ANDROID_HOME")
+                ?: System.getenv("ANDROID_SDK_ROOT")
+                ?: System.getenv("ANDROID_SDK_HOME")
+                ?: System.getenv("ANDROID_SDK")
+                ?: System.getenv("ANDROID")
+        }
+
+    private val androidUserHome: Path
+        get() {
+            if (System.getenv("ANDROID_USER_HOME") != null) {
+                return Paths.get(System.getenv("ANDROID_USER_HOME"))
+            }
+
+            return Paths.get(System.getProperty("user.home"), ".android")
+        }
+
+    /**
+     * Returns SDK versions that are used by AVDs present in the system.
+     */
+    val androidEmulatorSdkVersions: List<String>
+        get() {
+            val iniFiles = androidUserHome.resolve("avd").toFile()
+                .listFiles { file -> file.extension == "ini" }
+                ?.map { it } ?: emptyList()
+
+            val versions = iniFiles
+                .mapNotNull { iniFile -> iniFile.readLines().firstOrNull { it.startsWith("target=") } }
+                .map { line -> line.split('=') }
+                .filter { lineParts -> lineParts.size == 2 }
+                .map { lineParts -> lineParts[1] }
+                .distinct()
+                .toList()
+
+            return versions
+        }
+
     /**
      * @return Path to java compatible android cmdline-tools
      */
     fun requireCommandLineTools(tool: String): File {
-        val androidHome = EnvUtils.androidHome()
+        val androidHome = androidHome
             ?: throw CliError("Could not detect Android home environment variable is not set. Ensure that either ANDROID_HOME or ANDROID_SDK_ROOT is set.")
 
         val javaVersion = EnvUtils.getJavaVersion()
@@ -41,8 +81,8 @@ object AndroidEnvUtils {
             )
     }
 
-    private fun getRecommendedToolsVersion(): String? {
-        return when(EnvUtils.getJavaVersion()) {
+    private fun getRecommendedToolsVersion(): String {
+        return when (EnvUtils.getJavaVersion()) {
             8 -> "8.0"
             11 -> "10.0"
             17 -> "11.0"
@@ -51,14 +91,14 @@ object AndroidEnvUtils {
     }
 
     private fun findCompatibleCommandLineTool(tool: String): File? {
-        val path = File(EnvUtils.androidHome(), "cmdline-tools")
+        val path = File(androidHome, "cmdline-tools")
         return path.listFiles()
-            .filter { it.isDirectory && File(it, "/bin/$tool").exists() }
-            .filter { isCommandLineToolCompatible(File(it, "bin/$tool")) }
-            .sortedWith(compareBy<File> { it.name != "latest" }
+            ?.filter { it.isDirectory && File(it, "/bin/$tool").exists() }
+            ?.filter { isCommandLineToolCompatible(File(it, "bin/$tool")) }
+            ?.sortedWith(compareBy<File> { it.name != "latest" }
                 .thenByDescending { it.name.toDoubleOrNull() })
-            .map { File(it, "bin/$tool") }
-            .firstOrNull()
+            ?.map { File(it, "bin/$tool") }
+            ?.firstOrNull()
     }
 
     /**
@@ -83,12 +123,18 @@ object AndroidEnvUtils {
     fun parsePixelDevices(input: String): List<AvdDevice> {
         val pattern = "id: (\\d+) or \"(pixel.*?)\"\\n.*?Name: (.*?)\\n".toRegex()
         return pattern.findAll(input)
-            .map { matchResult -> AvdDevice(matchResult.groupValues[1], matchResult.groupValues[2], matchResult.groupValues[3]) }
+            .map { matchResult ->
+                AvdDevice(
+                    matchResult.groupValues[1],
+                    matchResult.groupValues[2],
+                    matchResult.groupValues[3]
+                )
+            }
             .toList()
     }
 
     fun requireEmulatorBinary(): File {
-        val androidHome = EnvUtils.androidHome()
+        val androidHome = androidHome
             ?: throw CliError("Could not detect Android home environment variable is not set. Ensure that either ANDROID_HOME or ANDROID_SDK_ROOT is set.")
         val firstChoice = File(androidHome, "emulator/emulator")
         val secondChoice = File(androidHome, "tools/emulator")
